+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 2.0.7 (Unreleased)
 ------------------
 
+- checkout tinymce language with - in addition to _
+  [vangheem]
 
 2.0.6 (2015-08-23)
 ------------------